rename worker file name
