+++ test/common/io/io_uring_worker_impl_test.cc
@@ -1,4 +1,4 @@
-#include "source/common/io/io_uring_worker.h"
+#include "source/common/io/io_uring_worker_impl.h"
 
 #include "test/mocks/event/mocks.h"
 #include "test/mocks/io/mocks.h"
